implemented session creation
EDDBK-154
