+++ src/modules/services/components/SessionEditor/index.js
@@ -1,4 +1,5 @@
 import template from './template.html'
+import { guessUnit } from '../../libs/sessionHelpers'
 
 export default function (AbstractEntityModalEditor, { mapState }) {
   return AbstractEntityModalEditor.extend({
@@ -23,9 +24,12 @@ export default function (AbstractEntityModalEditor, { mapState }) {
          * for the new session form.
          */
         model: {
+          id: null,
           sessionLength: null,
           price: null
         },
+
+        duration: null,
 
         timeUnits: [{
           title: this._('mins'),
@@ -59,27 +63,19 @@ export default function (AbstractEntityModalEditor, { mapState }) {
         entityModel: 'oneSession'
       }),
     },
+    mounted () {
+      this.$on('seed', () => {
+        const guessedUnit = guessUnit(this.model.sessionLength) || 0
+        this.sessionTimeUnit = this.timeUnits[guessedUnit].seconds
+        this.duration = Number(this.model.sessionLength) / this.sessionTimeUnit || null
+      })
+    },
     methods: {
-      /**
-       * Normalize session length information by transforming session duration
-       * in selected time units to duration in seconds.
-       *
-       * @param {SessionLength} sessionLength Session length to normalize.
-       *
-       * @return {SessionLength} Normalized session length.
-       */
-      normalizeSessionLength (sessionLength) {
-        const sessionDuration = this.sessionTimeUnit * sessionLength.sessionLength
-        return {
-          sessionLength: sessionDuration,
-          price: sessionLength.price
-        }
-      },
-
       /**
        * Validate current session and if everything is fine, save it.
        */
       saveSession () {
+        this.model.sessionLength = this.duration * this.sessionTimeUnit
         this.saveItem()
         // this.$validator.validateAll().then((result) => {
         //   if (!result) {
