+++ src/modules/services/components/Sessions/template.html
@@ -28,7 +28,7 @@
                         </div>
                         <div class="rich-row__column rich-row__actions">
                             <div class="button-group">
-                                <div class="button">Edit</div>
+                                <div class="button" @click="editSession(r.item)">Edit</div>
                                 <div class="button button-icon" @click="r.remove(r.item)"><span class="dashicons dashicons-trash"></span></div>
                             </div>
                         </div>
